Final Commit
Python script used to web scrape the details from adverts on Autotrader.com. The URL used in the script is already set up to filter results to listings valued between £500 - £1,000, within 5 miles of a specified postcode, and excluding cars that have been written off. Script writes the results to an output CSV file.